backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
kaggle==1.6.14
wbdata==1.0.0
ucimlrepo==0.0.7
requests==2.32.3
openpyxl==3.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
